Cleaned up ICS parsing logic, it is a bit more efficient now
